+++ src/elf.rs
@@ -52,10 +52,6 @@ impl<'file> Elf<'file> {
         })
     }
 
-    pub fn main_fn_address(&self) -> u32 {
-        self.symbols.main_fn_address
-    }
-
     pub fn program_uses_heap(&self) -> bool {
         self.symbols.program_uses_heap
     }
@@ -176,14 +172,12 @@ fn extract_debug_frame<'file>(elf: &ObjectFile<'file>) -> anyhow::Result<DebugFr
 }
 
 struct Symbols {
-    main_fn_address: u32,
     program_uses_heap: bool,
     reset_fn_range: Range<u32>,
     rtt_buffer_address: Option<u32>,
 }
 
 fn extract_symbols(elf: &ObjectFile, reset_fn_address: u32) -> anyhow::Result<Symbols> {
-    let mut main_fn_address = None;
     let mut program_uses_heap = false;
     let mut reset_symbols = Vec::new();
     let mut rtt_buffer_address = None;
@@ -196,7 +190,6 @@ fn extract_symbols(elf: &ObjectFile, reset_fn_address: u32) -> anyhow::Result<Sy
 
         let address = symbol.address().try_into().expect("expected 32-bit ELF");
         match name {
-            "main" => main_fn_address = Some(cortexm::clear_thumb_bit(address)),
             "_SEGGER_RTT" => rtt_buffer_address = Some(address),
             "__rust_alloc" | "__rg_alloc" | "__rdl_alloc" | "malloc" if !program_uses_heap => {
                 log::debug!("symbol `{}` indicates heap is in use", name);
@@ -211,7 +204,6 @@ fn extract_symbols(elf: &ObjectFile, reset_fn_address: u32) -> anyhow::Result<Sy
         }
     }
 
-    let main_fn_address = main_fn_address.ok_or(anyhow!("`main` symbol not found"))?;
     let reset_fn_range = {
         if reset_symbols.len() == 1 {
             let reset = reset_symbols.remove(0);
@@ -226,7 +218,6 @@ fn extract_symbols(elf: &ObjectFile, reset_fn_address: u32) -> anyhow::Result<Sy
     };
 
     Ok(Symbols {
-        main_fn_address,
         program_uses_heap,
         reset_fn_range,
         rtt_buffer_address,